bld: clean up old tests and add new ones for io
