--- tests/data/csv/mismatch_delimiter.csv
@@ -1,2 +1,0 @@
-Album;Artist;AlbumId;Track;Date#sid
-"a","b","","c","2020-12-10"
