--- tests/data/csv/no_username.csv
@@ -1,1 +1,0 @@
-Artist;Album;AlbumId;Track;Date#
