--- tests/data/csv/sample.csv
@@ -1,14 +1,0 @@
-Artist;Album;AlbumId;Track;Date#lazulinoother
-"Lana Del Rey";"Ultraviolence";"060d1168-7ea0-4290-9b05-ea3d23b14966";"Shades of Cool";"1594535082000"
-"Lana Del Rey";"Ultraviolence";"060d1168-7ea0-4290-9b05-ea3d23b14966";"Shades of Cool";"1594535254000"
-"Lana Del Rey";"Ultraviolence";"060d1168-7ea0-4290-9b05-ea3d23b14966";"Brooklyn Baby";"1594535609000"
-"Lana Del Rey";"Ultraviolence";"060d1168-7ea0-4290-9b05-ea3d23b14966";"West Coast";"1594535866000"
-"Lana Del Rey";"Ultraviolence";"060d1168-7ea0-4290-9b05-ea3d23b14966";"Sad Girl";"1594536184000"
-"Cigarettes After Sex";"I.";"50b51f4b-48ae-4f67-b94b-fe2f1aa87fb9";"Dreaming of You";"1594536486000"
-"Cigarettes After Sex";"I.";"50b51f4b-48ae-4f67-b94b-fe2f1aa87fb9";"Starry Eyes";"1594536692000"
-"Cigarettes After Sex";"Cigarettes After Sex";"0c68495e-485d-4976-bb85-438249649d08";"Sunsetz";"1594537250000"
-"Cigarettes After Sex";"Cigarettes After Sex";"0c68495e-485d-4976-bb85-438249649d08";"Flash";"1594537524000"
-"Cigarettes After Sex";"Cigarettes After Sex";"0c68495e-485d-4976-bb85-438249649d08";"Truly";"1594537767000"
-"Lana Del Rey";"Ultraviolence";"060d1168-7ea0-4290-9b05-ea3d23b14966";"Black Beauty";"1594538110000"
-"Lana Del Rey";"Ultraviolence";"060d1168-7ea0-4290-9b05-ea3d23b14966";"Guns and Roses";"1594538380000"
-"Lana Del Rey";"Ultraviolence";"060d1168-7ea0-4290-9b05-ea3d23b14966";"Florida Kilos";"1594538637000"
