--- tests/data/csv/valid_data.csv
@@ -1,2 +1,0 @@
-Artist;Album;AlbumId;Track;Date#lazulinoother
-"LDR";"UV";"a06";"Shades of Cool";"1594535082000"
