--- tests/data/csv/wrong_header.csv
@@ -1,2 +1,0 @@
-Artist;Album;AlbumId;Track;Date
-"ar";"alb";"";"tr";"2025-08-12"
