--- tests/data/json/canonical_json.json
@@ -1,1 +1,0 @@
-{"meta": {"username": "lazulinoother", "tz": "Asia/Kolkata", "num_scrobbles": 6, "date_range": {"start": "2020-07-12T11:54:42+05:30", "end": "2020-07-12T12:18:06+05:30"}, "source": "lastfmstats.com", "duration_present": false, "memory.fm_version": "0.4.3.dev5", "schema_version": 2}, "scrobbles": [{"timestamp": "2020-07-12T00:00:00+0530", "track": "Shades of Cool", "artist": "Lana Del Rey", "album": "Ultraviolence"}, {"timestamp": "2020-07-12T00:00:00+0530", "track": "Shades of Cool", "artist": "Lana Del Rey", "album": "Ultraviolence"}, {"timestamp": "2020-07-12T00:00:00+0530", "track": "Brooklyn Baby", "artist": "Lana Del Rey", "album": "Ultraviolence"}, {"timestamp": "2020-07-12T00:00:00+0530", "track": "West Coast", "artist": "Lana Del Rey", "album": "Ultraviolence"}, {"timestamp": "2020-07-13T00:00:00+0530", "track": "Sad Girl", "artist": "Lana Del Rey", "album": "Ultraviolence"}, {"timestamp": "2020-07-14T00:00:00+0530", "track": "Dreaming of You", "artist": "Cigarettes After Sex", "album": "I."}]}
