--- tests/data/json/wrong_column_2.json
@@ -1,1 +1,0 @@
-{"username": "lazulinoother", "column_2": ["Do you miss me, Miss Misery", "Like you say you do?"]}
