--- tests/data/test_to_json.json
@@ -1,1 +1,0 @@
-{"meta": {"username": "sid", "tz": "Europe/Berlin", "num_scrobbles": 1, "date_range": {"start": "2025-09-13T09:35:41+02:00", "end": "2025-09-13T09:35:41+02:00"}, "source": "lastfmstats.com", "memory.fm_version": "0.1.0.dev0", "schema_version": 1}, "scrobbles": [{"timestamp": "2025-09-13T09:35:41+0200", "track": "Days of Candy", "artist": "Beach House", "album": "Depression Cherry"}]}
